requirements.txt
fastapi
uvicorn[standard]
yt-dlp
deepgram-sdk[async]